#28. If a post userId is not equal to the session userId redirect a user to the posts-list page

+++ pages/post/delete-post.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 include '../set-project-root.php';
@@ -6,7 +7,16 @@ include PROJECT_ROOT . '/src/bootstrap.php';
 include '../private-page.php';
 
 use App\Controller\PostController;
+use App\Repository\PostRepository;
+use App\Service\DatabaseConnector;
 
+$db = DatabaseConnector::getDatabaseConnection();
+$postRepository = new PostRepository($db);
+$post = $postRepository->getPostById((int) $_GET['id']);
+if ($_SESSION['id'] !== $post['user_id']) {
+    header('Location: /pages/post/posts-list.php');
+    exit;
+}
 $isFailed = PostController::deletePost();
 ?>
